# docs/verification/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Global Soil Explorer API",
    "description": "REST API exposing Harmonized World Soil Database (HWSD) v2.0 dataset spatial lookup and database queries.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/soil": {
      "get": {
        "summary": "Query Soil Observation",
        "description": "Retrieves vertical soil profile properties, texture shares, and taxonomic classifications at the specified latitude and longitude coordinates.",
        "operationId": "get_soil_v1_soil_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Geographic latitude coordinate in decimal degrees (-90 to 90).",
              "example": 52.0,
              "title": "Latitude"
            },
            "description": "Geographic latitude coordinate in decimal degrees (-90 to 90)."
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Geographic longitude coordinate in decimal degrees (-180 to 180).",
              "example": 10.0,
              "title": "Longitude"
            },
            "description": "Geographic longitude coordinate in decimal degrees (-180 to 180)."
          }
        ],
        "responses": {
          "200": {
            "description": "Coordinates resolved and observation returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoilObservationSchema"
                }
              }
            }
          },
          "204": {
            "description": "Coordinates resolved to a void/ocean area (no observation)."
          },
          "400": {
            "description": "Geographic coordinates are invalid or out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "Underlying spatial index or database unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected application failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/soil": {
      "get": {
        "summary": "Query Soil Observation",
        "description": "Retrieves vertical soil profile properties, texture shares, and taxonomic classifications at the specified latitude and longitude coordinates.",
        "operationId": "get_soil_soil_get",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Geographic latitude coordinate in decimal degrees (-90 to 90).",
              "example": 52.0,
              "title": "Latitude"
            },
            "description": "Geographic latitude coordinate in decimal degrees (-90 to 90)."
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Geographic longitude coordinate in decimal degrees (-180 to 180).",
              "example": 10.0,
              "title": "Longitude"
            },
            "description": "Geographic longitude coordinate in decimal degrees (-180 to 180)."
          }
        ],
        "responses": {
          "200": {
            "description": "Coordinates resolved and observation returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoilObservationSchema"
                }
              }
            }
          },
          "204": {
            "description": "Coordinates resolved to a void/ocean area (no observation)."
          },
          "400": {
            "description": "Geographic coordinates are invalid or out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "Underlying spatial index or database unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected application failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CoordinateSchema": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude",
            "description": "Latitude of the coordinate in decimal degrees.",
            "example": 52.0
          },
          "longitude": {
            "type": "number",
            "title": "Longitude",
            "description": "Longitude of the coordinate in decimal degrees.",
            "example": 10.0
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "CoordinateSchema",
        "description": "Geographic coordinate schema."
      },
      "DatasetMetadataCodes": {
        "properties": {
          "coverage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage",
            "description": "Raw coverage source identifier code.",
            "example": 1
          }
        },
        "type": "object",
        "title": "DatasetMetadataCodes",
        "description": "Raw metadata codes."
      },
      "DatasetMetadataSchema": {
        "properties": {
          "coverage_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Description",
            "description": "Scientific description of the source map coverage.",
            "example": "FAO/UNESCO soil map of the world"
          },
          "library": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library",
            "description": "WRB library identifier name.",
            "example": "HWSD"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Active database or data source description.",
            "example": "HWSD v2.0 Database"
          },
          "dataset_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version",
            "description": "Active dataset version indicator.",
            "example": "v2.0"
          },
          "reference_identifiers": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Identifiers",
            "description": "Scientific reference identifiers (e.g., Soil Mapping Unit IDs).",
            "example": [
              [
                "SMU_ID",
                "10221"
              ]
            ]
          },
          "codes": {
            "$ref": "#/components/schemas/DatasetMetadataCodes",
            "description": "Grouped raw metadata codes."
          }
        },
        "type": "object",
        "required": [
          "codes"
        ],
        "title": "DatasetMetadataSchema",
        "description": "Database source provenance and dataset metadata."
      },
      "EnvironmentalContextSchema": {
        "properties": {
          "koppen_climate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Koppen Climate",
            "description": "Koppen-Geiger climate classification code.",
            "example": "C"
          }
        },
        "type": "object",
        "title": "EnvironmentalContextSchema",
        "description": "Climatic and environmental context of the soil site."
      },
      "ErrorResponseSchema": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Detailed explanation of the error.",
            "example": "Latitude must be a numeric value."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponseSchema",
        "description": "Standardized API error message response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HydrologicContextSchema": {
        "properties": {
          "drainage_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drainage Description",
            "description": "Resolved scientific description of natural soil drainage.",
            "example": "Well drained"
          },
          "water_regime_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Regime Description",
            "description": "Resolved scientific description of water regime characteristics."
          },
          "impermeable_layer_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impermeable Layer Description",
            "description": "Resolved scientific description of impermeable layer depth constraint."
          },
          "codes": {
            "$ref": "#/components/schemas/HydrologicContextCodes",
            "description": "Grouped raw hydrologic codes."
          }
        },
        "type": "object",
        "required": [
          "codes"
        ],
        "title": "HydrologicContextSchema",
        "description": "Hydrologic soil attributes schema representing scientific descriptions."
      },
...
      "SoilClassificationSchema": {
        "properties": {
          "taxonomy_standard": {
            "type": "string",
            "title": "Taxonomy Standard",
            "description": "Taxonomic classification standard.",
            "example": "WRB 2022"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name",
            "description": "Taxonomic class description name.",
            "example": "Luvisols"
          },
          "wrb4_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wrb4 Name",
            "description": "WRB 4th Edition classification name."
          },
          "wrb2_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wrb2 Name",
            "description": "WRB 2nd Edition classification name."
          },
          "fao90_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fao90 Name",
            "description": "FAO 1990 classification name."
          },
          "wrb_phase_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wrb Phase Name",
            "description": "WRB soil phase name description."
          },
          "national_classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "National Classification",
            "description": "National soil classification system descriptor."
          },
          "codes": {
            "$ref": "#/components/schemas/SoilClassificationCodes",
            "description": "Grouped raw taxonomic codes."
          }
        },
        "type": "object",
        "required": [
          "taxonomy_standard",
          "class_name",
          "codes"
        ],
        "title": "SoilClassificationSchema",
        "description": "Taxonomic soil classification schema representing standard scientific values."
      },
      "SoilLayerSchema": {
        "properties": {
          "top_depth_cm": {
            "type": "number",
            "title": "Top Depth Cm",
            "description": "Depth to top of layer in centimeters.",
            "example": 0.0
          },
          "bottom_depth_cm": {
            "type": "number",
            "title": "Bottom Depth Cm",
            "description": "Depth to bottom of layer in centimeters.",
            "example": 30.0
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/SoilPropertySchema"
            },
            "type": "array",
            "title": "Properties",
            "description": "Chemical and physical property values list (backward compatible).",
            "deprecated": true
          },
          "texture": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SoilTextureSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Texture codes for this layer depth."
          },
          "measurements": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LayerMeasurementsSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Consolidated chemical and physical measurement groups."
          }
        },
        "type": "object",
        "required": [
          "top_depth_cm",
          "bottom_depth_cm",
          "properties"
        ],
        "title": "SoilLayerSchema",
        "description": "Vertical soil layer schema interval."
      },
      "SoilObservationSchema": {
        "properties": {
          "coordinate": {
            "$ref": "#/components/schemas/CoordinateSchema",
            "description": "The query coordinate associated with this observation."
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/SoilProfileSchema"
            },
            "type": "array",
            "title": "Profiles",
            "description": "The list of soil profiles matching this spatial location."
          },
          "environmental_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnvironmentalContextSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Environmental and climate descriptors of the site."
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetMetadataSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data provenance, standards, and dataset metadata."
          }
        },
        "type": "object",
        "required": [
          "coordinate",
          "profiles"
        ],
        "title": "SoilObservationSchema",
        "description": "Scientific observation at location coordinates."
      },
      "SoilProfileSchema": {
        "properties": {
          "layers": {
            "items": {
              "$ref": "#/components/schemas/SoilLayerSchema"
            },
            "type": "array",
            "title": "Layers",
            "description": "Stack of vertical soil layers ordered by depth."
          },
          "classification": {
            "$ref": "#/components/schemas/SoilClassificationSchema",
            "description": "Taxonomic classification details."
          },
          "composition_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Composition Share",
            "description": "Percentage composition of this profile in the spatial unit.",
            "example": 80.0
          },
          "hydrologic_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HydrologicContextSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hydrologic attributes of this profile."
          },
          "land_limitations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LandLimitationsSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Agronomic and growth limitation ratings."
          },
          "sequence_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence Index",
            "description": "Rank of this profile in the overall mapping unit.",
            "example": 1
          }
        },
        "type": "object",
        "required": [
          "layers",
          "classification"
        ],
        "title": "SoilProfileSchema",
        "description": "Multi-layer vertical soil profile schema."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}